6 changes to project - Added an about panel for users who open in the GUI - Added Xcode release and debug configurations - Added 64bit support - Fixed localization for resources - Fixed types and pointers through out source - Renamed COPYING to LICENSE

+++ src/CDControl/CDBubbleControl.h
@@ -29,7 +29,7 @@
 // We really ought to stick this in a proper NSColor category
 + (NSColor *) colorFromHex:(NSString *) hexValue alpha:(float)alpha;
 
-- (NSColor *) _colorForBubble:(int)i fromKey:(NSString *)key alpha:(float)alpha;
+- (NSColor *) _colorForBubble:(unsigned long)i fromKey:(NSString *)key alpha:(float)alpha;
 
 // returns an NSArray of NSImage's or nil if there's only one.
 - (NSArray *) _iconImages;
