+++ src/CDControl/CDThreeButtonControl/CDRadioControl.m
@@ -140,7 +140,7 @@
     // Setup the control
     NSArray *items = [[[NSArray alloc] init] autorelease];
 	items = [options optValues:@"items"];
-    NSInteger selected = -1;
+    unsigned long selected = -1;
     NSArray *disabled = [[[NSArray alloc] init] autorelease];
 
 
@@ -156,7 +156,7 @@
     int matrixPrecedence = 0;
     
     // Set default number of columns
-    NSInteger columns = 1;
+    unsigned long columns = 1;
     // Set specified number of columns
     if ([options hasOpt:@"columns"]) {
         columns = [[options optValue:@"columns"] intValue];
@@ -166,7 +166,7 @@
     }
     
     // Set default number of rows
-    NSInteger rows = 1;
+    unsigned long rows = 1;
     // Set specified number of rows
     if ([options hasOpt:@"rows"]) {
         rows = [[options optValue:@"rows"] intValue];
@@ -190,7 +190,7 @@
     NSMutableArray * controls = [[[NSMutableArray alloc] init] autorelease];
     
     // Create the control for each item
-    NSInteger currItem = 0;
+    unsigned long currItem = 0;
     NSEnumerator *en = [items objectEnumerator];
     float cellWidth = 0.0;
     id obj;
@@ -221,8 +221,8 @@
     
     // Populate the matrix
     currItem = 0;
-    for (int currColumn = 0; currColumn <= columns - 1; currColumn++) {
-        for (int currRow = 0; currRow <= rows - 1; currRow++) {
+    for (unsigned long currColumn = 0; currColumn <= columns - 1; currColumn++) {
+        for (unsigned long currRow = 0; currRow <= rows - 1; currRow++) {
             if (currItem <= [items count] - 1) {
                 NSButtonCell * cell = [controls objectAtIndex:currItem];
                 [controlMatrix putCell:cell atRow:currRow column:currColumn];
